requirements.txt
Django==1.6.5
MySQL-python==1.2.5
South==0.8.4
django-allauth==0.17.0
ipython==2.1.0
jsonfield==0.9.20
numpy==1.8.1
oauthlib==0.6.3
pyparsing
python-dateutil==2.2
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
six==1.10.0
tornado==3.2.2
wsgiref==0.1.2
simplejson==3.6.5
boto==2.34.0